requirements.txt
Flask
Flask-Security
Flask-Script
Flask-SQLAlchemy
Flask-Migrate
Flask-Basicauth
Flask-Cache
Flask-Elasticsearch
mysql-python
python-memcached